created custom hook for product filters
