+++ src/pages/ProductList.jsx
@@ -1,8 +1,8 @@
-import { useState, useEffect, useContext } from 'react';
-import { useNavigate, createSearchParams } from 'react-router-dom';
+import { useContext } from 'react';
+// import { useNavigate, createSearchParams } from 'react-router-dom';
 import Spinner from 'react-bootstrap/Spinner';
-import { useProductCategories } from '../utils/hooks/useProductCategories';
-import { useAllProducts } from '../utils/hooks/useAllProducts';
+// import { useProductCategories } from '../utils/hooks/useProductCategories';
+// import { useAllProducts } from '../utils/hooks/useAllProducts';
 import { GlobalContext } from '../context/GlobalContext';
 import FilteredProducts from '../components/FilteredProducts';
 import ProductFilters from '../components/ProductFilters';
@@ -11,80 +11,13 @@ import { PageContainer } from '../components/styles/PageContainer.styled';
 import { SectionTitle } from '../components/styles/SectionTitle.styled';
 import { ProductListSection } from '../components/styles/ProductListSection.styled';
 import { SpinnerContainer } from '../components/styles/SpinnerContainer.styled';
+import { useFilters } from '../utils/hooks/useFilters';
 
 const ProductList = () => {
   const { selectedCategory } = useContext(GlobalContext);
-  const navigate = useNavigate();
+  const filtersData = useFilters(selectedCategory);
 
-  const { data: allProductsData, isLoading: productsAreLoading } =
-    useAllProducts();
-
-  const { data: categoriesData, isLoading: categoriesAreLoading } =
-    useProductCategories(selectedCategory);
-
-  const [filteredProducts, setFilteredProducts] = useState([]);
-
-  const [productCategories, setProductCategories] = useState([]);
-
-  const getActiveCategories = () => {
-    const activeCategories = productCategories.filter(
-      (category) => category.active
-    );
-    const activeCategoriesIds = activeCategories.map((category) => category.id);
-    return activeCategoriesIds;
-  };
-
-  const filterProducts = (activeCategoriesIds) => {
-    const newFilteredProducts =
-      allProductsData.results &&
-      allProductsData.results.filter((product) =>
-        activeCategoriesIds.includes(product.data.category.id)
-      );
-    setFilteredProducts(newFilteredProducts);
-  };
-
-  const toggleCategory = (id) => {
-    const modifiedCategories = productCategories.map((category) =>
-      category.id === id ? { ...category, active: !category.active } : category
-    );
-    setProductCategories(modifiedCategories);
-  };
-
-  const clearAllFilters = () => {
-    const clearedCategories = productCategories.map((category) => {
-      return { ...category, active: false };
-    });
-    setProductCategories(clearedCategories);
-  };
-
-  useEffect(() => {
-    const activeCategoriesIds = getActiveCategories();
-
-    if (activeCategoriesIds.length === 0) {
-      setFilteredProducts(
-        allProductsData.results ? [...allProductsData.results] : []
-      );
-      navigate('/products');
-      return;
-    }
-    filterProducts(activeCategoriesIds);
-    const params = { category: activeCategoriesIds };
-    navigate({
-      pathname: '/products',
-      search: `?${createSearchParams(params)}`,
-    });
-  }, [productCategories]);
-
-  useEffect(() => {
-    setFilteredProducts(
-      allProductsData.results ? [...allProductsData.results] : []
-    );
-    setProductCategories(
-      categoriesData.results ? [...categoriesData.results] : []
-    );
-  }, [allProductsData, categoriesData]);
-
-  if (productsAreLoading || categoriesAreLoading) {
+  if (filtersData.productsAreLoading || filtersData.categoriesAreLoading) {
     return (
       <SpinnerContainer>
         <Spinner animation="border" variant="danger" />
@@ -97,11 +30,11 @@ const ProductList = () => {
       <SectionTitle>Product List</SectionTitle>
       <ProductListSection>
         <ProductFilters
-          toggleCategory={toggleCategory}
-          productCategories={productCategories}
-          clearAllFilters={clearAllFilters}
+          toggleCategory={filtersData.toggleCategory}
+          productCategories={filtersData.productCategories}
+          clearAllFilters={filtersData.clearAllFilters}
         />
-        <FilteredProducts filteredProducts={filteredProducts} />
+        <FilteredProducts filteredProducts={filtersData.filteredProducts} />
         <PaginationButtons />
       </ProductListSection>
     </PageContainer>
